+++ src/lib/fgetline/fsize.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 
-int estiva_fsize(FILE *fp)
+long estiva_fsize(FILE *fp)
 {
   long i=0;
   while( EOF != fgetc(fp)) i++;
